Part 1 almost done, read discription
All changes are also commented, and I clarified some ToDos! But here's a rundown.

Broke up the parsing and maping stuff into an indexing function. Was easier to call and seemed more readable, don't be afraid to put it back in main / make changes.

Added a step where before assigning words and vectors to the main_map, the parsed data is stored in a list of allSentences. This should make indexing new data within the same session much easier to implement, as well as fufilled the requirements for printing sentences from the terminal.

Only thing I wasnt sure about was passing things like allSentences and main_map as parameters to functions, and if that was logical. Don't think what I implemented affects run time at all.

All that's left to do before we're set on Part 1 submission is add the commands to the help file. And I'll do that within the next few hours.